Add very basic card profile generation outline
