+++ config/dataOpts.json
@@ -10,6 +10,7 @@
             "0x40": "Divine",
             "0x80": "Laugh"
         },
+        "banlist": "https://raw.githubusercontent.com/Ygoproco/Live2017Links/master/lflist.conf",
         "categories": {
             "0x1": "Destroy Spell/Trap",
             "0x2": "Destroy Monster",
@@ -131,6 +132,7 @@
             "0x40": "Divine",
             "0x80": "Laugh"
         },
+        "banlist": "https://raw.githubusercontent.com/Ygoproco/Live2017Links/master/lflist.conf",
         "categories": {
             "0x1": "Destroy Spell/Trap",
             "0x2": "Destroy Monster",
